Atualização controllers e entidades Evento/Ingresso

+++ src/main/java/com/conectaarena/repository/EventoRepository.java
@@ -1,17 +1,14 @@
 package com.conectaarena.repository;
 
 import com.conectaarena.model.Evento;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
+import org.springframework.data.jpa.repository.JpaRepository;//interface com métodos prontos de banco de dados
+import org.springframework.stereotype.Repository; //anotacao repository
 import java.time.LocalDateTime;
-import java.util.List;
+import java.util.List;//pra poder mostrar varios eventos
 
 @Repository
-public interface EventoRepository extends JpaRepository<Evento, Long> {
-
-    // buscar eventos com data maior que a atual, ordenados pela mais próxima
+public interface EventoRepository extends JpaRepository<Evento, Integer>{
+    //métodos de consulta , ver a data dos proximos eventos e ver por categoria
     List<Evento> findByDataAfterOrderByDataAsc(LocalDateTime data);
-
-    // busca por categoria e data futura
     List<Evento> findByCategoriaAndDataAfterOrderByDataAsc(String categoria, LocalDateTime data);
 }
